Add detection of zettelkasten path
It is a bother that my zettelkasten is located in different places
within each system.

As a fix I have set the zshrc to store the location in a
~/.config/zettelkasten-path. If the file does not exist, the
zettelkasten will be searched for and the file is created. Then the
ZETTELKASTEN variable is loaded with the value in the file.